requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.1.2
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.1.2
fastapi==0.115.2
fastapi-cache2==0.2.2
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.2
idna==3.10
Jinja2==3.1.4
kombu==5.4.2
magic-filter==1.0.12
makefun==1.15.6
Mako==1.3.5
MarkupSafe==3.0.1
multidict==6.1.0
pendulum==3.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.10
pwdlib==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
redis==4.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.16.0
